Move stuff in right package (i.e., the default package for actions etc.)
