+++ alchemist/alchemist-incarnation-scafi/src/main/scala/it/unibo/alchemist/model/implementations/conditions/ScafiComputationalRoundComplete.scala
@@ -6,10 +6,10 @@
  * GNU General Public License, with a linking exception,
  * as described in the file LICENSE in the Alchemist distribution's top directory.
  */
-package it.unibo.alchemist.implementations.conditions
+package it.unibo.alchemist.model.implementations.conditions
 
-import it.unibo.alchemist.implementations.nodes.ScafiNode
-import it.unibo.alchemist.implementations.actions.RunScafiProgram
+import it.unibo.alchemist.model.implementations.nodes.ScafiNode
+import it.unibo.alchemist.model.implementations.actions.RunScafiProgram
 import it.unibo.alchemist.model.ScafiIncarnationUtils
 import it.unibo.alchemist.model.implementations.conditions.AbstractCondition
 import it.unibo.alchemist.model.interfaces.{Condition, Context, Node, Reaction}
