+++ alchemist/alchemist-incarnation-scafi/src/main/scala/it/unibo/alchemist/model/implementations/nodes/NodeManager.scala
@@ -5,7 +5,7 @@
  * GNU General Public License, with a linking exception,
  * as described in the file LICENSE in the Alchemist distribution's top directory.
  */
-package it.unibo.alchemist.implementations.nodes
+package it.unibo.alchemist.model.implementations.nodes
 
 import it.unibo.alchemist.model.implementations.molecules.SimpleMolecule
 import it.unibo.alchemist.model.interfaces.{Node}
